# backend/openapi_snapshot.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "My Darrin API",
                 "version":  "0.1.0"
             },
    "paths":  {
                  "/api/v1/auth/register":  {
                                                "post":  {
                                                             "tags":  [
                                                                          "Authentication",
                                                                          "Auth"
                                                                      ],
                                                             "summary":  "Register",
                                                             "operationId":  "register_api_v1_auth_register_post",
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/RegisterRequest"
                                                                                                                                     }
                                                                                                                      }
                                                                                             },
                                                                                 "required":  true
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "Successful Response",
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {

                                                                                                                                               }
                                                                                                                                }
                                                                                                       }
                                                                                       },
                                                                               "422":  {
                                                                                           "description":  "Validation Error",
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         }
                                            },
                  "/api/v1/auth/login":  {
                                             "post":  {
                                                          "tags":  [
                                                                       "Authentication",
                                                                       "Auth"
                                                                   ],
                                                          "summary":  "Login",
                                                          "operationId":  "login_api_v1_auth_login_post",
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/LoginRequest"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Successful Response",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {

                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                            "422":  {
                                                                                        "description":  "Validation Error",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      }
                                         },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "LoginRequest":  {
                                                            "properties":  {
                                                                               "email":  {
                                                                                             "type":  "string",
                                                                                             "format":  "email",
                                                                                             "title":  "Email"
                                                                                         },
                                                                               "password":  {
                                                                                                "type":  "string",
                                                                                                "title":  "Password"
                                                                                            }
                                                                           },
                                                            "type":  "object",
                                                            "required":  [
                                                                             "email",
                                                                             "password"
                                                                         ],
                                                            "title":  "LoginRequest"
                                                        },
                                       "RegisterRequest":  {
                                                               "properties":  {
                                                                                  "email":  {
                                                                                                "type":  "string",
                                                                                                "format":  "email",
                                                                                                "title":  "Email"
                                                                                            },
                                                                                  "password":  {
                                                                                                   "type":  "string",
                                                                                                   "title":  "Password"
                                                                                               }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "email",
                                                                                "password"
                                                                            ],
                                                               "title":  "RegisterRequest"
                                                           },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           },
                                                                                  "input":  {
                                                                                                "title":  "Input"
                                                                                            },
                                                                                  "ctx":  {
                                                                                              "type":  "object",
                                                                                              "title":  "Context"
                                                                                          }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   },
                       "securitySchemes":  {
                                               "HTTPBearer":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              }
                                           }
                   }
}